ai-assistant/requirements.txt
# AI/ML Core
awesome-deepseek==0.1.0
requests==2.31.0
python-dotenv==1.0.0

# Web & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0

# Kafka
confluent-kafka==2.2.0
kafka-python==2.0.2

# Utilities
numpy==1.24.3
pandas==2.0.3